optimize test structure, clean
