+++ tests/e2e/test_chaperone.js
@@ -10,87 +10,27 @@ const http_servers = require('../setup_http_server.js');
 const setup = require("../setup_envoy.js");
 const setup_conductor = require("../setup_conductor.js");
 const { Codec } = require('@holo-host/cryptolib');
-const installedAppIds = yaml.load(fs.readFileSync('./script/app-config.yml'));
-const { resetTmp, delay } = require("../utils")
+const { create_page, fetchServiceloggerCellId, setupServiceLoggerSettings, PageTestUtils, envoy_mode_map, resetTmp, delay } = require("../utils")
 const msgpack = require('@msgpack/msgpack');
-
-// NOTE: the test app servicelogger installed_app_id is hard-coded, but intended to mirror our standardized installed_app_id naming pattern for each servicelogger instance (ie:`${hostedAppHha}::servicelogger`)
-const HOSTED_APP_SERVICELOGGER_INSTALLED_APP_ID = installedAppIds[0].app_name;
 
 const INVALID_JOINING_CODE = msgpack.encode('failing joining code').toString('base64')
 const SUCCESSFUL_JOINING_CODE = msgpack.encode('joining code').toString('base64')
-
-let browser;
-
-async function create_page(url) {
-  const page = await browser.newPage();
-
-  log.info("Go to: %s", url);
-  await page.goto(url, {
-    "waitUntil": "networkidle0"
-  });
-
-  return page;
-}
-
-class PageTestUtils {
-  constructor(page) {
-    this.logPageErrors = () => page.on('pageerror', async error => {
-      if (error instanceof Error) {
-        log.silly(error.message);
-      } else
-        log.silly(error);
-    });
-
-    this.describeJsHandleLogs = () => page.on('console', async msg => {
-      const args = await Promise.all(msg.args().map(arg => this.describeJsHandle(arg)))
-        .catch(error => console.log(error.message));
-      console.log(args);
-    });
-
-    this.describeJsHandle = (jsHandle) => {
-      return jsHandle.executionContext().evaluate(arg => {
-        if (arg instanceof Error)
-          return arg.message;
-        else
-          return arg;
-      }, jsHandle);
-    };
-  }
-}
 
 // NB: The 'host_agent_id' *is not* in the holohash format as it is a holo host pubkey (as generated from the hpos-seed)
 const host_agent_id = 'd5xbtnrazkxx8wjxqum7c77qj919pl2agrqd3j2mmxm62vd3k'
 
 log.info("Host Agent ID: %s", host_agent_id);
 
-const envoy_mode_map = {
-  production: 0,
-  develop: 1,
-}
-
 // Note: All envoyOpts.dnas will be registered via admin interface with the paths provided here
 const envoyOpts = {
   mode: envoy_mode_map.develop,
 }
 
-const getHostAgentKey = async (appClient) => {
-  const appInfo = await appClient.appInfo({
-    installed_app_id: HOSTED_APP_SERVICELOGGER_INSTALLED_APP_ID
-  });
-  const agentPubKey = appInfo.cell_data[0].cell_id[1];
-  return {
-    decoded: agentPubKey,
-    encoded: Codec.AgentId.encode(agentPubKey)
-  }
-}
 const REGISTERED_HAPP_HASH = "uhCkkCQHxC8aG3v3qwD_5Velo1IHE1RdxEr9-tuNSK15u73m1LPOo"
 
 describe("Server", () => {
-  let envoy;
-  let server;
-  let http_ctrls, http_url;
-  let registered_agent;
+  let envoy, server, browser
+  let http_ctrls, http_url, page;
 
   before(async function() {
     this.timeout(100_000);
@@ -117,6 +57,27 @@ describe("Server", () => {
     browser = await puppeteer.launch();
     log.debug("Setup config: %s", http_ctrls.ports);
     http_url = `http://localhost:${http_ctrls.ports.chaperone}`;
+  
+    const page_url = `${http_url}/html/chaperone.html`
+    page = await create_page(page_url, browser);
+    const pageTestUtils = new PageTestUtils(page)
+
+    pageTestUtils.logPageErrors();
+    pageTestUtils.describeJsHandleLogs();
+
+    await page.exposeFunction('delay', delay)
+
+    // Set logger settings for hosted app (in real word scenario - will be done when host installs app):
+    try {
+      const servicelogger_cell_id = await fetchServiceloggerCellId(envoy.hcc_clients.app);
+      console.log("Found servicelogger cell_id: %s", servicelogger_cell_id);
+      // NOTE: The host settings must be set prior to creating a service activity log with servicelogger (eg: when making a zome call from web client)
+      const logger_settings = await setupServiceLoggerSettings(envoy.hcc_clients.app, servicelogger_cell_id);
+      console.log("happ service preferences set in servicelogger as: %s", logger_settings);
+    } catch (err) {
+      console.log(typeof err.stack, err.stack.toString());
+      throw err;
+    }
   });
 
   after(async () => {
@@ -142,7 +103,6 @@ describe("Server", () => {
 
   it('should fail to sign up without wormhole', async function () {
     this.timeout(30_000)
-
     const { Client: RPCWebsocketClient } = require('rpc-websockets')
 
     const agentId = 'uhCAkgHic-Y_Y1C-o9MvNW9KwnqGTNDxyQLjxnL2hETY6BXgONqlT'
@@ -163,7 +123,7 @@ describe("Server", () => {
     expect(response).deep.equal({
       name: 'HoloError',
       message:
-        'HoloError: Error: CONDUCTOR CALL ERROR: {"type":"internal_error","data":"Conductor returned an error while using a ConductorApi: GenesisFailed { errors: [ConductorApiError(WorkflowError(SourceChainError(KeystoreError(LairError(Other(OtherError(\\"unexpected: ErrorResponse { msg_id: 9, message: \\\\\\"Failed to fulfill hosted signing request: \\\\\\\\\\\\\'Failed to get signature from Chaperone\\\\\\\\\\\\\'\\\\\\" }\\")))))))] }"}'
+        'HoloError: Error: CONDUCTOR CALL ERROR: {"type":"internal_error","data":"Conductor returned an error while using a ConductorApi: GenesisFailed { errors: [ConductorApiError(WorkflowError(SourceChainError(KeystoreError(LairError(Other(OtherError(\\"unexpected: ErrorResponse { msg_id: 11, message: \\\\\\"Failed to fulfill hosted signing request: \\\\\\\\\\\\\'Failed to get signature from Chaperone\\\\\\\\\\\\\'\\\\\\" }\\")))))))] }"}'
     })
     const closedPromise = new Promise(resolve => client.once("close", resolve))
     client.close()
@@ -172,58 +132,12 @@ describe("Server", () => {
 
   it("should sign-in and make a zome function call", async function() {
     this.timeout(300_000);
-
     try {
-      const page_url = `${http_url}/html/chaperone.html`
-      const page = await create_page(page_url);
-      const pageTestUtils = new PageTestUtils(page)
-
-      pageTestUtils.logPageErrors();
-      pageTestUtils.describeJsHandleLogs();
-
-      await page.exposeFunction('fetchServiceloggerCellId', async () => {
-        let serviceloggerCellId;
-        try {
-          // REMINDER: there is one servicelogger instance per installed hosted app, each with their own installed_app_id
-          const serviceloggerAppInfo = await envoy.hcc_clients.app.appInfo({
-            installed_app_id: HOSTED_APP_SERVICELOGGER_INSTALLED_APP_ID
-          });
-          serviceloggerCellId = serviceloggerAppInfo.cell_data[0].cell_id;
-        } catch (error) {
-          throw new Error(JSON.stringify(error));
-        }
-        return serviceloggerCellId;
-      });
-
-      // Note: the host must set servicelogger settings prior to any activity logs being issued (otherwise, the activity log call will fail).
-      await page.exposeFunction('setupServiceLoggerSettings', async (servicelogger_cell_id) => {
-        const settings = {
-          // Note: for the purposes of simplifying the test, the host is also the provider
-          provider_pubkey: Codec.AgentId.encode(servicelogger_cell_id[1]),
-          max_fuel_before_invoice: 3,
-          price_compute: 1,
-          price_storage: 1,
-          price_bandwidth: 1,
-          max_time_before_invoice: [604800, 0]
-        }
-        let logger_settings;
-        logger_settings = await envoy.hcc_clients.app.callZome({
-          // Note: Cell ID content MUST BE passed in as a Byte Buffer, not a u8int Byte Array
-          cell_id: [Buffer.from(servicelogger_cell_id[0]), Buffer.from(servicelogger_cell_id[1])],
-          zome_name: 'service',
-          fn_name: 'set_logger_settings',
-          payload: settings,
-          cap: null,
-          provenance: Buffer.from(servicelogger_cell_id[1])
-        });
-        return logger_settings;
-      });
-
       await page.exposeFunction('encodeHhaHash', (type, buf) => {
         const hhaBuffer = Buffer.from(buf);
         return Codec.HoloHash.encode(type, hhaBuffer);
       });
-      const { responseOne, responseTwo } = await page.evaluate(async function (host_agent_id, registered_agent, registered_happ_hash, SUCCESSFUL_JOINING_CODE) {
+      const { responseOne, responseTwo } = await page.evaluate(async function (host_agent_id, registered_happ_hash, joiningCode) {
         console.log("Registered Happ Hash: %s", registered_happ_hash);
 
         const client = new Chaperone({
@@ -244,7 +158,7 @@ describe("Server", () => {
         client.skip_assign_host = true;
 
         await client.ready(200_000);
-        await client.signUp("alice.test.1@holo.host", "Passw0rd!", SUCCESSFUL_JOINING_CODE);
+        await client.signUp("alice.test.1@holo.host", "Passw0rd!", joiningCode);
         console.log("Finished sign-up for agent: %s", client.agent_id);
         if (client.anonymous === true) {
           throw new Error("Client did not sign-in")
@@ -253,18 +167,6 @@ describe("Server", () => {
           throw new Error(`Unexpected Agent ID: ${client.agent_id}`)
         }
 
-        // Set logger settings for hosted app (in real word scenario - will be done when host installs app):
-        try {
-          const servicelogger_cell_id = await window.fetchServiceloggerCellId();
-          console.log("Found servicelogger cell_id: %s", servicelogger_cell_id);
-
-          // NOTE: The host settings must be set prior to creating a service activity log with servicelogger (eg: when making a zome call from web client)...
-          const logger_settings = await window.setupServiceLoggerSettings(servicelogger_cell_id);
-          console.log("happ service preferences set in servicelogger as: %s", logger_settings);
-        } catch (err) {
-          console.log(typeof err.stack, err.stack.toString());
-          throw err;
-        }
         let responseOne, responseTwo;
         try {
           responseOne = await client.callZomeFunction(`test`, "test", "pass_obj", {'value': "This is the returned value"});
@@ -274,13 +176,6 @@ describe("Server", () => {
           throw err
         }
 
-        function delay(t, val) {
-          return new Promise(function(resolve) {
-            setTimeout(function() {
-              resolve(val);
-            }, t);
-          });
-        }
         // Delay is added so that the zomeCall has time to finish all the signing required
         //and by signing out too soon it would not be able to get all the signature its needs and the test would fail
         await delay(15000);
@@ -289,7 +184,7 @@ describe("Server", () => {
 
         // Test for second agent on same host
         // NEED TO FIX
-        await client.signUp("bob.test.1@holo.host", "Passw0rd!");
+        await client.signUp("bob.test.1@holo.host", "Passw0rd!", joiningCode);
         console.log("Finished sign-up for agent: %s", client.agent_id);
         if (client.anonymous === true) {
           throw new Error("Client did not sign-in")
@@ -304,7 +199,7 @@ describe("Server", () => {
           responseOne,
           responseTwo
         }
-      }, host_agent_id, registered_agent, REGISTERED_HAPP_HASH, SUCCESSFUL_JOINING_CODE);
+      }, host_agent_id, REGISTERED_HAPP_HASH, SUCCESSFUL_JOINING_CODE);
 
       log.info("Completed evaluation: %s", responseOne);
       log.info("Completed evaluation: %s", responseTwo);
